+++ python/gnomon_utils/__init__.py
@@ -1,2 +1,0 @@
-from .gnomonDataDriverMongo import gnomonDataDriverMongo
-from .gnomonPlugin import load_plugin_group, corePlugin, visualizationPlugin, available_plugins, serialize